ROMFonts: Build fix
To avoid confusion on the round trip of DOS->Unix->RISC OS don't keep the /Base0 and /Default encodings files with that name. For example on Unix this ends up as ".Base0" and ".Default" which are hidden files.
Adjust MyDesk macro to perform this transform.
Escape the '$' symbols hidden in various strings.
Binary inspected manually, looks the same.

Version 0.77. Tagged as 'ROMFonts-0_77'