Env: Define Install$Path in addition to Install$Dir
The !Common script will now define a macro variable for Install$Path
which contains <Install$Dir>. which means install targets can now be
specified in the form "Install:foo" rather than "<Install$Dir>.foo".

This is especially handy because any path that has a / in it can end
up being munged from UNIX to RISC OS form, even if it was a RISC OS
path in the first place. But one thing that prevents the munging is
if the path has a : in it.

Long story short, you can now do stuff like this:

    INSTDIR=Install:DDE.AcornC/C++.Library

which was hopelessly broken prior to this change.

Admin:
  Requires mkdir 1.04 for the full goodness. Will be put into the
  main Library component at some point soon.

Version 1.29. Tagged as 'Env-1_29'