FSLock: Change to use objasm
Previously this component relied on BASIC to assemble the module (and patch in the default password etc).
Instead, use objasm so we can pull in all the normal system defines and make use of it's stricter checking.
A new tool "scrambler" performs the operation of patching in the default password.
Notes -
 This is a literal translation of former 'FSLockSRC', bugs and all.
 Put several unused files in the attic.
 Tested in an IOMD ROM.

Version 1.23. Tagged as 'FSLock-1_23'