DesktopBoot: Correction to Boot$ReadOnly evaluation logic
BootVars was not accounting for a locked drive (with FSLock) when working out whether to set Boot$ReadOnly.
This meant that a locked drive would unhelpfully create a small RAM disc with a default setup in it rather than running the normal boot sequence (albeit locked against changes to everything except $.Public and !Scrap).
Tested on RISC OS 4.02.

Version 1.21. Tagged as 'Boot-1_21'