ScrnSetup: Look for territory specific messages before falling back to en-gb
To build the screen saver 'Info' menu this plugin reaches into the screensaver to get the purpose/author/version fields from its messages, however was not respecting any territory selection.
This change emulates how Toolbox_Initialise would look and is essentially a repeat of ticket #367.
Also - the return registers for MessageTrans_FileInfo were wrong, R0 and R2 (not R1 and R2) but since the flags were never used, this is easier to delete.

Version 0.32. Tagged as 'ScrnSetup-0_32'