PDModules: The case of the mysteriously vanishing text
When trying to print text written in system font using DrawFile to
render it, via PDriverDP, the system font text would vanish while
outline fonts would be printed correctly. Creating a drawing in !Draw
and embedding the result in (for example) !TechWriter would trigger the
problem, as it uses DrawFile_Render to handle the embedded Draw area.

Turns out to be a quirk of the way DrawFile renders system font text. It
creates a temporary MODE0 8x8 sprite on the stack, switches output to
it, uses OS_WriteC in VDU 4 mode to plot one character, switches output
back (which at that point would be to PDriverDP which has output
switched in order to make the bit image to print). This process repeats
for all characters in the text object.

When the kernel switches output to a sprite/mask you inherit the
CursorFlags that were active at the point of the switch - you're free to
change them if desired, but the important part is they're inherited.
Unfortunately PDriverDP, in set_sprite_output_state, was turning off
scroll protect which meant when DrawFile plotted its 1 character to
convert, the kernel plotted it as requested, but then the cursor reached
the end of the line (which is only 1 long) and the sprite was scrolled
up to the new line - making it blank again.

PDriverPS would show the text correctly, since it doesn't make the image
by redirecting output to a bitmap, hence leaves the default CursorFlags
in play.

This could have been sidestepped if DrawFile had generated the text a
different way, for example using OS_SpriteOp 51, or by locally turning
on scroll protection in the CursorFlags for its sprite.

The PRM is a bit contradictory on the matter. PRM3-589 states that the
cursor movement control bits are set to &40, then PRM3-590 says that
"This is designed to be as similar as possible to the state set up by
the window manager". The latter is not the case, as the Wimp messes with
CursorFlags in 4 places
* initvdustring1 which preserves the state of scroll protect
* writedirnormal which preserves the state of scroll protect
* writedirreverse which preserves the state of scroll protect
* scrollprot which turns on scroll protect
indeed, for it to work in the non printing case DrawFile is already relying on
this behaviour.

So, set_sprite_output_state now enables scroll protect like the Wimp,
and the case of the mysteriously vanishing text is closed.

Version 4.60. Tagged as 'PDModules-4_60'