Econet: Import of SetStation sources
Recreated with the help of !SID, this is the utility previously supplied with AUN to set a machine's station number. Note, it pokes IOC and has various other 26 bit delights in it.
Produces almost the same binary, with the exception of LDRB rn,[rd] which AAsm used to create LDRB rn,[rd],#0 for,
Not tagged.