Patience: Fix some minor inconsistencies !Sprites22 were using a new mode word, when the 256 colour/no palette/mask sprites can be described by numbered mode 28. The Filer showed now sprite pre RISC OS 3.50.
The program uses Wimp_ResizeIcon and the nested form of Wimp_OpenWindow so should declare '380' to Wimp_Initialise.
The game window menu didn't line up right in system font, which requires the correct number of spaces in the messages.
The clipping box for some of the options in the Choices dialogue weren't big enough in system font.

Tested on an A5000 with Wimp 3.98 loaded.

Version 0.74. Tagged as 'Patience-0_74'