SquashApp: Fix some minor inconsistencies
The blunt check for UtilityModule 3.60 didn't make sense given it's followed by a check for CLib 3.99, since RISC OS 3.60 has CLib 4.79 in ROM.
Careful inspection of the C code shows it doesn't use any exciting SWIs and the calls to RISCOS_Lib are similarly simple (as !Squash only has a menu and save box as a user interface).
Therefore, allow it to use RISC OS 3.10, but require a CLib that supports APCS-32 since that's what all the disc components will be built with.

Tested on an A5000 with CLib 5.83.

Version 0.55. Tagged as 'Squash-0_55'