HAL_Tungsten: Minor HAL API corrections
HAL_HardwareInfo - as word 0 specified that no IOMD or VIDC was present, don't waste time reporting anything in word 2.
HAL_PhysInfo - report the ROM range as an inclusive range for op 2, same as the RAM is for op 1.
Requires Kernel 4.79.2.248 to get the corresponding ROM range change.

Version 0.54. Tagged as 'Tungsten-0_54'