HAL_OMAP4: Enable L2 cache in HAL_Init
On the OMAP4 we're in a privileged but unsecure mode, so to do the initial PL310 enable we call an SMC sunction embedded in the OMAP's onboard mask ROM to do the deed.
Also, correct the range of SDRAM hinted to the kernel - this looks like a copy and paste mistake from the Tungsten HAL.

Version 0.40. Tagged as 'OMAP4-0_40'