ADFSFiler: Fix for weird text in floppy format menu
The menu icon block for the various native format floppy discs was writing the length and validation string pointers the wrong way round. By chance, there's now a vaguely valid validation string character at the address of the length of the "ADFS 1600K (F+)" entry, which made is into a string of 'o's for some reason.
Also, corrected the width calculation for half a system font character at either end - not that relevant as the Wimp's calculated it for years.
Tested briefly in an IOMD ROM, menu looks right.

Version 1.08. Tagged as 'ADFSFiler-1_08'