squeeze: Fix wrapping of modules into squeezed apps
When presented with a module, squeeze will squeeze it and tack on a header which does OS_Module 11 on it when unsqueezed.
However, the header size (bytes) was being added to a pointer (of type int) leading to the header being loaded 4x too high, overwriting the heap - size now adjusted.

Also remove the assumption that clock() counts in cs, so that the times come out right on a cross compiled Unix system.

Version 5.14. Tagged as 'squeeze-5_14'