PatchesPatch: Fix to unsqueeze recent apps requiring patching
Squeeze 5.11 and later inserts a dummy signature to convince ROL's UnSqueezeAIF to let it pass, however the Unsqueeze command inside !Patch didn't get updated to expect that and merrily branched into the dummy one, ending up with an abort at whatever was in R4.

Having found a signature match, check at +20 to see if this was a dummy (in line with the logic of UnSqzAIF-0_09). This version can afford to be more simplistic since it never wants to run the resulting application, it doesn't really matter if the StrongARM cache flush code is right or not, so there's only 2 to choose from.
Update the Squeeze binary to squeeze-5_13.
Tested with a few binaries from 5.08 and some from 5.13.
Fixes ticket #403.

Version 1.38. Tagged as 'Patch-1_38'