Pogo: Numerous screensaver/setup improvements
Reworked the sprite plotting to not leave a trail of flickering pogo stick animation on the screen, this was caused by confusion about OS units mostly and writing into the wrong screen bank.
Replaced the squash code with some that's sensible - this now uses a squash file exactly as output by !Squash rather than a homebrew format. The screensaver app was also supplied with both the squashed and not squashed versions of the sprite just for extra comedy.
Changed to using common functions in "saver.c" rather than local ones.
There's no setup app for this saver as there's nothing to configure.
Other admin changes:
* Makefiles recreated from fragments
* Version number now auto inserted from VersionNum
* Messages and common code "saver.c" sync'd with other screensaver/setup apps
* Removed references to Filer Configure from RMEnsure error messages
* Fix null pointer dereference when -openat switch has too few arguments
* Symbolic names used instead of magic numbers where possible
Tested on RISC OS 4.02.

Version 0.04. Tagged as 'Pogo-0_04'