Swarm: Numerous screensaver/setup improvements
The default swarm speed in the code didn't match the default in the setup app. Additionally, the default was 0, which on modern CPUs is impossibly fast flying for bees - so when using the 'Try' option in the screen setup plugin the result didn't make sense nor match what the dialogue showed.
UK run files moved into UK resources.
Other admin changes:
* Makefiles recreated from fragments
* Version number now auto inserted from VersionNum
* Messages and common code "saver.c" sync'd with other screensaver/setup apps
* Removed references to Filer Configure from RMEnsure error messages
* Removed never used info from setup app messages, ensuring they're never out of date by deleting them
* Removed non internationalised "Screensave setup" from error box titles
* Load settings from Choices: write settings through <Choices$Write>
* Fix null pointer dereference when -openat switch has too few arguments
* Symbolic names used instead of magic numbers where possible
Tested on RISC OS 4.02.

Version 0.05. Tagged as 'Swarm-0_05'