Econet: Add support for RAM variant
Now we use the shared makefiles & accept MessageTrans is present, we get a RAM variant with merged messages almost for free. Add the supporting ResourceFS register/deregister code, this removes the need for a separate EconetMsgs module.
Tested on a RISC OS 3.70 Risc PC with AEH60 card, browsing a server & copying files to and fro.

Version 5.76. Tagged as 'Econet-5_76'