BuildSys: Extend resources phase to handle more than just Messages
Using the InstViaRG script CModule will now collate a list of files (using the syntax for the InstRes script employed by the CApp shared makefile) for resources, rather than the previous Messages file only.
The list is specified in INSTRES_FILES, or if for some reason the standalone and ROM versions need to be different INSTRAM_FILES and INSTROM_FILES (cf. INSTAPP_FILES in CApp).
Shock new documentation for CModule added.

Version 6.70. Tagged as 'BuildSys-6_70'