MakeModes: Fix for NULL pointer dereference
When first selecting a monitor there are no modes generated yet, so don't try to check the line rate or frame rate for a NULL mode. Do check it when the monitor is subsequently changed (as the modes exist then).
Enlarge the text box for "1MB VRAM" so it's not clipped in Homerton 12pt.
Tested on a Pi with zero page protection.

Version 0.31. Tagged as 'MakeModes-0_31'