DiscSetup: Raise RAM disc clamp when possible
If the 'PMP' extensions for dynamic areas are available, allow the numberrange gadget to choose a larger RAM disc, and if not clamp at 128MB as before.
Use unsigned types to hold memory amounts, and calculate in pages where possible to avoid overflow/signedness issues above 2GB.
As at time of writing RAMFS is still being worked on, this plugin doesn't respect that, so if you choose a size larger than RAMFS supports it'll choose the lower of the two.

Tested briefly on a Risc PC with 256MB RAM with and without PMP dynamic areas.

Version 0.11. Tagged as 'DiscSetup-0_11'