JCompMod: Add support for CompressJPEG_Comment tag injection
You can now annotate the JPEGs you create with this handy pocket sized SWI. The maximum comment length of 65000 comes from the 'wrjpgcom' utility from the IJG library of the same vintage.
Tested briefly with a modified version of ChangeFSI to put my name in, and cross check the result on http://regex.info/exif.cgi

Version 0.08. Tagged as 'JCompMod-0_08'