ColourDbox: Fix for NULL pointer dereference when 'Default' title is selected
Contrary to the Toolbox manual, ColourDbox was taking a zero "max title length" field as the cue to substitute the title, followed by setting the title string pointer to NULL, trying to strlen() it, and ending up with the default from the Res file for this object. The code to look up the default title in the Messages was never called.
Changed to follow the docs: a NULL title string pointer uses the default and no significance is attached to the max buffer length in that case.
Messages file (which is now used as a result) changed to have the same default title as before.

Thanks to Nick Roberts for spotting & testing this.

Version 0.21. Tagged as 'ColourDbox-0_21'