MakeModes: Various fixes
Choosing 'VESA' from the iconbar menu now ticks it (was broken due to capitalising the directory name, and the menu handling code needed lower case 'n').
Fix data abort when creating an MDF from nothing, this is a very longstanding bug, caused by trying to copy the contents of the icon_g_type to the global_monitor_type which was never used outside the function, nor did it have any underlying storage (was NULL).
Tidying
  Line 1993: Make add_monitor_to_list() static.
  Line 259: Remove global SWIregs since it's only used once in a function in a different source file.
  Line 2225: Too many spaces on the query message.
  Line 783/789: Comment which INKEY values are being used.

Version 0.32. Tagged as 'MakeModes-0_32'