EtherY: Debug correction, and fix for mbuf loop on fragmented packet
eymodule.c: Correction to debug message typo.
glue.c: On fast machines if you're unlucky while sending out a fragmented packet which has been provided as a list of mbufs length > 1, having sent the first chain in the list by using CAT() the chain is free'd but without cropping the hdr off which now points to a free'd mbuf. An incoming packet just at that moment would likely recycle the mbuf and the result is a circular linked list which MbufManager can't process. Unlink hdr before freeing to break the chain.

Version 0.63. Tagged as 'EtherY-0_63'