PortableHAL: Add sensor reading SWI
New Portable_ReadSensor is a general front end to sensors that might be of concern to the Portable module.
Starter for 10 is temperature, and we support reading the CPU die temperature (from a new CPUClk entry point, since faster=botter they're physically related parameters) and the BMU charger. Each supports multiple units, so in future multicore temperatures can be reported.
New error message in the resources, plus enable tokenisation of the Messages file since it's used via a buffered MessageTrans_ErrorLookup call.

Tested briefly on a HAL with fake temperature sensor returning a hardwired kelvin value.

Version 0.76. Tagged as 'PortableHAL-0_76'