InetSetup: Fade the 'Run RouteD' option when the RouteD module is missing
Ticking the option would cause an error during booting, since the module cannot be loaded.
Removed the default option text from the Res file, as it's not language specific, it's a CLI switch looked for by the module. Strictly speaking making it lowercase (like the "default" accepted as a netmask) is an incompatible change as disassembling a 1993 copy of the RouteD module shows it uses strcmp() on "DEFAULT", but since that wont run on anything even as recent as a StrongARM it's probably better to note that any new port should be using a case insensitive switch check.

Version 0.55. Tagged as 'Source-0_55'