SprExtend: Fix corrupt R0 during JPEG plot
The JPEG_PlotScaled SWI (and friends) was pulling R0 at the wrong moment, so corrupting it. This manifested as "Incomplete or corrupt JPEG" when trying to plot a JPEG via the PostScript driver for example (which expects R0 preserved for no error).
Looks like ~5 year old bug, presumably not many people printing JPEGs to PostScript, or all suffering in silence.
Tested on a Risc PC - printed a picture from !Draw to postscript, then PDF'd it.

Version 1.69. Tagged as 'SprExtend-1_69'