SoftLoad: Update to support ARMv7
Unlike ARMv5 and earlier, we can't rely on the final 'MMU off & jump' to be safely in the pipeline, fancy ARMs can do speculative execution and branch prediction ending up trying to run code that's not yet there followed by trying to execute exception vectors that also aren't there. Solve this by flat mapping the 'MMU off & jump' code like we do the ROM itself and jump to the copy in logical address space that's in the same place as in physical space.
softload.c:
  Express memory amounts using size_t so checks on machines with > 2GB don't think they have -ve memory amounts.
  Export log2phys() for the assembler to use.
  Use bool to hold booleans.
  Trim some unused/redundant defines.
startnew.s:
  Lookup the physical address of AppSpaceStart ready for L1PT flat mapping.
  Lookup the MMU_Changing ARMOp from the kernel rather than mess about with local copies (requires RISC OS 5.23 or later).
  Add code to skip the XScale/StrongARM bits and bobs based on the MIDR.
  Fix bug setting U bit in L1PT on ARMs that don't support it.
  Leave xsf alone when switching processor mode in CPSR.
  Use Push/Pull macros.
Tested on Titanium, and regression tested on a Risc PC softloading RISC OS 5.

Version 1.21. Tagged as 'SoftLoad-1_21'